Add files via upload
Participating in the Zindi UmojaHack2023 Cryptojacking Detection Challenge (INTERMEDIATE) was an exciting experience for my team and me. We focused on cybersecurity, data analysis, and machine learning to fight against cryptojacking. After putting in a lot of effort and using our skills, we achieved an impressive ranking of 3 out of 247 teams. This shows that our approach, algorithms, and models were effective in finding and stopping unauthorized crypto mining. Our success in the challenge makes us feel more confident and shows that we know a lot about cybersecurity. We are eager to keep working on making the digital world safer, and we're thankful for the chance to be part of this important competition.